Split up `ext::secondary` file
